LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.get.unaffected.worker-expected.txt


PASS getImageData() is not affected by context state 
